requirements.txt
python-dotenv==1.0.0
openai==1.12.0
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3 